requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
jmespath==1.0.1
pathspec==0.10.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
semantic-version==2.8.5
six==1.16.0
sqlparse==0.5.0
termcolor==1.1.0
urllib3==1.26.18
utils==1.0.2
wcwidth==0.1.9
